pyproject.toml
[project]
name = "wagtail-shiki"
authors = [{name = "kawakin", email = "kawakin26@gmail.com"},]
description = "Add a code block to StreamField of Wagtail with Shiki Syntax highlighter."
dynamic = ["version"]
readme = "README.md"
requires-python = ">=3.9"
keywords = ["wagtail", "shiki", "cms", "syntax", "code", "highlighting", "highlighter"]
license = "License :: OSI Approved :: BSD License"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.1",
    "Framework :: Django :: 5.2",
    "Framework :: Wagtail :: 4",
    "Framework :: Wagtail :: 5",
    "Framework :: Wagtail :: 6",
    "Framework :: Wagtail :: 7",
]
dependencies = [
    "wagtail>=4",
]

[project.optional-dependencies]
dev = [
  "django-coverage-plugin",
  "ipython",
  "ruff",
  "pytest-coverage",
  "pytest-django",
]

[project.urls]
"Homepage" = "https://github.com/kawakin26/wagtail-shiki"
"Repository" = "https://github.com/kawakin26/wagtail-shiki"
"Documentation" = "https://github.com/kawakin26/wagtail-shiki"

[build-system]
requires = ["setuptools>=67", "setuptools_scm>=7", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
write_to = "wagtail_shiki/_version.py"

[tool.pytest.ini_options]
addopts = "--cov --cov-report=html"
python_files = "tests.py test_*.py"
DJANGO_SETTINGS_MODULE = "tests.settings"

[tool.coverage.run]
plugins = ["django_coverage_plugin"]
include = ["wagtail-shiki/*"]
